Update patch set 1
Patch Set 1: Verified+1

Depends on http://review.cyanogenmod.org/#/c/98958/

I tested reporting 0 cameras if /dev/video0 is not available, which prevents the TorchService crash.  However, you have to reboot after attaching a camera to get the camera HAL to recognize the camera that was attached.  This is not the ideal behavor.  With these two patches plug and play operation of the USB webcam works as it should.

Patch-set: 1