Removing chmod/chown call to dev/dx_sep_q1
There is a call to change permissions on a device driver that is not
reachable from other processes other than init and ueventd. These
useless device drivers should be stripped out of the kernel. This change
will handle removing the call to change permissions on them to
facilitate removing these devices out of all relevant kernels.

Bug: 33347297
Test: The device boots
Change-Id: I04b58d4551c4d6215833aadb1a51afe11b0c0bcf