Update patch set 1
Patch Set 1: Code-Review-2

> Uploaded patch set 1.

Don't merge yet, this is pending some testing with in-remote mic. We may actually be able to support it's audio routing in 8.1 (due to minimal  changes from 8.0), but as with previous versions, its likely 9.0 will break it permanently.

I also submitted a request for the remotecontrolservice PDK to Intel's OSS team, as the code comments mention it being pushed to OSS at some point in the future. Hopefully we get a response on that one.

Patch-set: 1